fix(GromovPolynomialGrowth): `HasPolynomialGrowth` should only quantify over positive `n` (#1259)
- fix `HasPolynomialGrowth` to quantify only over `n > 0`
- add some more API and tests to demonstrate the issue with `n = 0`: if
`G` is an infinite group, then no matter what `d` is, we can always find
`n` such that `C * n ^ d < GrowthFunction S n`. If `d > 0` then `n = 0`
trivially works. On the other hand, if `d = 0` then this reduces to
showing that `GrowthFunction S n` is unbounded which holds in the
infinite case.
- this generalises the counterexample of `Multiplicative ℤ` discovered
by AlphaProof to `GromovPolynomialGrowthTheorem`, since this is an
infinite finitely-generated nilpotent group.

---------

Co-authored-by: Calle Sönne <csonne@google.com>
Co-authored-by: Paul Lezeau <lezeau@google.com>
Co-authored-by: Yaël Dillies <yael.dillies@gmail.com>